Small email confirmation bug fix
If user had an unconfirmed email in the database, a further edit would
not hit the database until confirmation--not the experience I was
looking for.

If a user has a confirmed email in the database, it will stay in the
database until the new one has been confirmed.

Removed some debug print statements.